+++ package.js
@@ -27,7 +27,7 @@ function configure( api ){
     api.use( 'blaze-html-templates@2.0.0', 'client' );
     api.use( 'ecmascript' );
     api.use( 'less@4.0.0', 'client' );
-    api.use( 'pwix:i18n@1.0.0' );
+    api.use( 'pwix:i18n@1.5.2' );
     api.use( 'pwix:jquery-ui@1.0.1' );
     api.use( 'pwix:layout@1.3.0' );
     api.use( 'random', 'client' );