+++ package.js
@@ -28,7 +28,7 @@ function configure( api ){
     api.use( 'ecmascript' );
     api.use( 'less@4.0.0', 'client' );
     api.use( 'pwix:i18n@1.5.2' );
-    api.use( 'pwix:jquery-ui@1.0.1' );
+    api.use( 'pwix:jquery-ui@1.0.2' );
     api.use( 'pwix:layout@1.3.0' );
     api.use( 'random', 'client' );
     api.use( 'tmeasday:check-npm-versions@1.0.2', 'server' );